Add a function to move tiles in a row
The basic rules of the game are as follows:

* The game is played on a grid with numbered tiles.
* On every move all tiles slide in a chosen direction.
* Each tile slides in the chosen direction as far as possible until it's
stopped by either another tile or the edge of the grid.
* If two tiles of the same number collide, they are merged into a tile
with their total value.
* The resulting tile cannot merge with another tile again in the same
move.
* If a move causes three consecutive tiles of the same value to slide
together, only the two tiles farthest along the direction of motion will
combine.
* If four spaces in a row or column are filled with tiles of the same
value, a move parallel to that row/column will combine the first two and
last two.

For now we'll assume that each row/column of the grid will be
represented by a list of nils and integers, where an integer represents
a tile and a nil represents an empty space, ordered so that on every
move the tiles are moved to the beginning of the list.

This commit adds a function that sorts and merges tiles in a list
according to the rules above.