Place tile of value 2 in a random space when starting a game
When a game is started, the grid must be empty except for one tile of
value 2 placed at random.

The test creates a new game 5 time to verify that the tile is not placed
in the same location every time. 5 is random number and should be
increased if the test proves to be flaky.