Make it possible to restart the game in the UI
Once the game has ended (player left, game won or lost), it'll work like
this:

* Any user can restart the game.
* The first user to restart the game becomes player1.
* All remaining users can join the game.
* The first user to join becomes player2.
* All remaining users become spectators.

Once one of the players leave the game, it's also necessary to stop
monitoring the LiveView process of the other player, as otherwise, if
the game is restarted and that player doesn't join, when they leave the
new game will be stopped. I.e. we only want to monitor the players of
the current game.