Move validation logic to Muster.Game
Validating if a player can join a game or make a move is "business
logic" and it's always the same, regardless of how the game is played.
Therefore Muster.Game is a better place for it, then Muster.CurrentGame
which is just a way to provide a single "current" game and synchronise
updates to it.