Add a script to start the app
The script makes it possible to run the app via Docker or Elixir/Node on
the host. It tries to use Docker if possible.

If Docker is installed, it'll build an image for the app, install
dependencies and start Phoenix server on port 4000.

If Docker is not installed, it'll attempt to do the same on the host
directly.