Animate tiles
I wanted to animate:

* Appearance of a new tile on the grid
* Movement on tiles into the direction chosen by a player
* Disappearanace of tiles that are merged together

I also wanted to see if it'd be possible to do it by relying on LiveView
and CSS animations only.

It's possible to animate DOM nodes with CSS when they appear or when
their certain CSS attributes change, which has several implications:

* A tile needs to be represented by the same DOM node throughout the
entire game. Therefore it's necessary to introduce a `Tile` struct with
a unique ID (instead of using nested list of `nil`s / `integer`s to
represent the grid) and make `Grid` operate on those.
* LiveView needs to be prevented from adding a new DOM node for an
existing tile. It's not possible to control HTML diffing that LiveView
does explicitly, but to ensure that tile nodes remain in place after
creation (and only their attributes / contents are updated) implicitly
it's necessary to always render all tiles that have been added
(including the ones that have been merged, but making them hidden) and
in the same order. This means that over the course of the game quite a
few tiles that are not actually visible will be accumulated in the DOM,
which is not ideal at all and is an indicator that doing this kind of
thing using LiveView might not be a good idea.
* Position of elements in a CSS grid can't be animated, so animating
movement requires absolute positioning of tiles, which makes the styling
less flexible.